+++ app/models/datasource_scientific_domain.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class DatasourceScientificDomain < ApplicationRecord
-  belongs_to :datasource
-  belongs_to :scientific_domain
-
-  validates :datasource, presence: true
-  validates :scientific_domain, presence: true, uniqueness: { scope: :datasource_id }
-end