+++ app/models/persistent_identity_system.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class PersistentIdentitySystem < ApplicationRecord
-  belongs_to :datasource, foreign_key: "datasource_id", inverse_of: :persistent_identity_systems
+  belongs_to :datasource, foreign_key: "service_id", inverse_of: :persistent_identity_systems, class_name: "Service"
 
   belongs_to :entity_type, class_name: "Vocabulary::EntityType", optional: true
   has_many :persistent_identity_system_vocabularies, dependent: :destroy