+++ config/routes.rb
@@ -37,8 +37,6 @@ Rails.application.routes.draw do
     end
   end
 
-  resources :datasources, only: [:index, :show], constraints: { id: %r{[^/]+} }
-
   resource :comparisons, only: [:show, :destroy] do
     scope module: :comparisons do
       resource :services, only: [:create, :destroy]
@@ -82,7 +80,7 @@ Rails.application.routes.draw do
 
   resource :backoffice, only: :show
   namespace :backoffice do
-    resources :services, constraints: { id: %r{[^/]+} } do
+    resources :services, controller: "services", constraints: { id: %r{[^/]+} } do
       scope module: :services do
         resource :logo_preview, only: :show
         resources :offers
@@ -97,12 +95,6 @@ Rails.application.routes.draw do
     resources :categories
     resources :providers, constraints: { id: %r{[^/]+} }
     resources :platforms
-    resources :datasources, constraints: { id: %r{[^/]+} } do
-      scope module: :datasources do
-        resource :publish, only: :create
-        resource :draft, only: :create
-      end
-    end
     get "vocabularies", to: "vocabularies#index", type: "target_user", as: :vocabularies
     scope "/vocabularies" do
       resources :target_users, controller: "vocabularies", type: "target_user"
@@ -125,7 +117,7 @@ Rails.application.routes.draw do
       resources :research_steps, controller: "vocabularies", type: "research_step"
       resources :jurisdictions, controller: "vocabularies", type: "jurisdiction"
       resources :datasource_classifications, controller: "vocabularies", type: "datasource_classification"
-      resources :research_entity_types, controller: "vocabularies", type: "research_entity_type"
+      resources :entity_types, controller: "vocabularies", type: "entity_type"
       resources :entity_type_schemes, controller: "vocabularies", type: "entity_type_scheme"
       resources :product_access_policies, controller: "vocabularies", type: "product_access_policy"
       resources :bundle_goals, controller: "vocabularies", type: "bundle_goal"