+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Improvements for the UI of bundles (@jarekzet)
 - Interoperability guidelines import (@michal-szostak)
 
+### Changed
+- Datasources as services (@goreck888)
+
+### Fixed
+- REST data import token exchange (@goreck888)
+
 ## [3.47.3] 2023-04-03
 
 ### Fixed